requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
--find-links https://data.pyg.org/whl/torch-2.4.0+cu124.html

ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
#numpy==1.24.4
#scipy
#pandas==1.5.3    
#pip
pytest
python-dotenv
#scikit-learn
tqdm
typer
-e .

# PyTorch core
torch==2.4.0+cu124
torchvision==0.19.0+cu124
torchaudio==2.4.0+cu124

# —–– PyG low-level CUDA kerneli:
pyg_lib==0.4.0+pt24cu124
torch_scatter==2.1.2+pt24cu124
torch_sparse==0.6.18+pt24cu124
torch_cluster==1.6.3+pt24cu124
torch_spline_conv==1.2.2+pt24cu124

# —–– PyG high-level:
torch-geometric==2.3.1
torch-geometric-temporal==0.56.0
pytorch-lightning==2.2.4
# DVC for experiment tracking
dvc
dvclive
# YAML parsing
pyyaml